+++ lib/src/arithmetic/values/fixed_point_value.dart
@@ -1,4 +1,4 @@
-// Copyright (C) 2024-2025 Intel Corporation
+// Copyright (C) 2024-2026 Intel Corporation
 // SPDX-License-Identifier: BSD-3-Clause
 //
 // fixed_point_value.dart
@@ -40,14 +40,25 @@ class FixedPointValue implements Comparable<FixedPointValue> {
   /// Returns `true` if the number is negative.
   bool isNegative() => signed & (value[-1] == LogicValue.one);
 
-  /// Constructs [FixedPointValue] from [integer] and [fraction] values with a
-  /// [signed] option to interpret MSB of [integer] as sign bit with the
-  /// [integer] represented in twos-complement.
+  /// Constructs an unsigned [FixedPointValue] from [integer] and [fraction].
+  ///
+  /// Use [FixedPointValue.withSignedness] for new code.
+  @Deprecated('Use FixedPointValue.withSignedness instead.')
   factory FixedPointValue(
           {required LogicValue integer,
           required LogicValue fraction,
+          bool signed = false}) =>
+      FixedPointValue.withSignedness(
+          integer: integer, fraction: fraction, signed: signed);
+
+  /// Constructs [FixedPointValue] from [integer] and [fraction] values with a
+  /// [signed] option to interpret the MSB of [integer] as a two's-complement
+  /// sign bit.
+  factory FixedPointValue.withSignedness(
+          {required LogicValue integer,
+          required LogicValue fraction,
           bool signed = true}) =>
-      populator(
+      populatorWithSignedness(
               integerWidth: integer.width - (signed ? 1 : 0),
               fractionWidth: fraction.width,
               signed: signed)
@@ -56,16 +67,36 @@ class FixedPointValue implements Comparable<FixedPointValue> {
   /// Creates an unpopulated version of a [FixedPointValue], intended to be
   /// called with the [populator].
   @protected
-  FixedPointValue.uninitialized({this.signed = true});
+  @Deprecated('Use FixedPointValue.uninitializedWithSignedness instead.')
+  FixedPointValue.uninitialized({this.signed = false});
+
+  /// Creates an unpopulated [FixedPointValue] with the specified signedness.
+  @protected
+  FixedPointValue.uninitializedWithSignedness({this.signed = true});
 
   /// Creates a [FixedPointValuePopulator] with the provided [integerWidth]
   /// and [fractionWidth], which can then be used to complete construction of
   /// a [FixedPointValue] using population functions.
+  ///
+  /// Use [populatorWithSignedness] for new code.
+  @Deprecated('Use FixedPointValue.populatorWithSignedness instead.')
   static FixedPointValuePopulator populator(
           {required int integerWidth,
           required int fractionWidth,
+          bool signed = false}) =>
+      populatorWithSignedness(
+          integerWidth: integerWidth,
+          fractionWidth: fractionWidth,
+          signed: signed);
+
+  /// Creates a [FixedPointValuePopulator] with the provided widths and
+  /// signedness.
+  static FixedPointValuePopulator populatorWithSignedness(
+          {required int integerWidth,
+          required int fractionWidth,
           bool signed = true}) =>
-      FixedPointValuePopulator(FixedPointValue.uninitialized(signed: signed)
+      FixedPointValuePopulator(
+          FixedPointValue.uninitializedWithSignedness(signed: signed)
         ..integerWidth = integerWidth
         ..fractionWidth = fractionWidth);
 
@@ -77,7 +108,8 @@ class FixedPointValue implements Comparable<FixedPointValue> {
   /// of [FixedPointValue]s.
   @mustBeOverridden
   FixedPointValuePopulator clonePopulator() =>
-      FixedPointValuePopulator(FixedPointValue.uninitialized(signed: signed)
+      FixedPointValuePopulator(
+          FixedPointValue.uninitializedWithSignedness(signed: signed)
         ..integerWidth = integerWidth
         ..fractionWidth = fractionWidth);
 
@@ -95,14 +127,14 @@ class FixedPointValue implements Comparable<FixedPointValue> {
     final s = signed | other.signed;
     final m = max(integerWidth, other.integerWidth);
     final n = max(fractionWidth, other.fractionWidth);
-    final val1 =
-        FixedPointValue.populator(integerWidth: m, fractionWidth: n, signed: s)
-            .widen(this)
-            .value;
-    final val2 =
-        FixedPointValue.populator(integerWidth: m, fractionWidth: n, signed: s)
-            .widen(other)
-            .value;
+    final val1 = FixedPointValue.populatorWithSignedness(
+            integerWidth: m, fractionWidth: n, signed: s)
+        .widen(this)
+        .value;
+    final val2 = FixedPointValue.populatorWithSignedness(
+            integerWidth: m, fractionWidth: n, signed: s)
+        .widen(other)
+        .value;
     final comp = val1.compareTo(val2);
     if (comp == 0) {
       return comp;
@@ -264,15 +296,15 @@ class FixedPointValue implements Comparable<FixedPointValue> {
     final s = signed | other.signed;
     final nr = max(fractionWidth, other.fractionWidth);
     final mr = max(integerWidth, other.integerWidth) + 1;
-    final val1 = FixedPointValue.populator(
+    final val1 = FixedPointValue.populatorWithSignedness(
             integerWidth: mr, fractionWidth: nr, signed: s)
         .widen(this)
         .value;
-    final val2 = FixedPointValue.populator(
+    final val2 = FixedPointValue.populatorWithSignedness(
             integerWidth: mr, fractionWidth: nr, signed: s)
         .widen(other)
         .value;
-    return FixedPointValue.populator(
+    return FixedPointValue.populatorWithSignedness(
             integerWidth: mr, fractionWidth: nr, signed: s)
         .ofLogicValue(val1 + val2);
   }
@@ -286,17 +318,19 @@ class FixedPointValue implements Comparable<FixedPointValue> {
       throw RohdHclException('Inputs must be valid.');
     }
     // Subtraction can produce a negative result regardless of the operands'
-    // signedness, so the result is always signed (matching the
-    // [FixedPointValue.populator] default).
+    // signedness, so the result is always signed.
     final nr = max(fractionWidth, other.fractionWidth);
     final mr = max(integerWidth, other.integerWidth) + 1;
-    final val1 = FixedPointValue.populator(integerWidth: mr, fractionWidth: nr)
+    final val1 = FixedPointValue.populatorWithSignedness(
+            integerWidth: mr, fractionWidth: nr)
         .widen(this)
         .value;
-    final val2 = FixedPointValue.populator(integerWidth: mr, fractionWidth: nr)
+    final val2 = FixedPointValue.populatorWithSignedness(
+            integerWidth: mr, fractionWidth: nr)
         .widen(other)
         .value;
-    return FixedPointValue.populator(integerWidth: mr, fractionWidth: nr)
+    return FixedPointValue.populatorWithSignedness(
+            integerWidth: mr, fractionWidth: nr)
         .ofLogicValue(val1 - val2);
   }
 
@@ -313,19 +347,19 @@ class FixedPointValue implements Comparable<FixedPointValue> {
         : integerWidth + other.integerWidth;
     final nr = fractionWidth + other.fractionWidth;
     final tr = mr + nr;
-    final val1 = FixedPointValue.populator(
+    final val1 = FixedPointValue.populatorWithSignedness(
             integerWidth: tr - fractionWidth,
             fractionWidth: fractionWidth,
             signed: s)
         .widen(this)
         .value;
-    final val2 = FixedPointValue.populator(
+    final val2 = FixedPointValue.populatorWithSignedness(
             integerWidth: tr - other.fractionWidth,
             fractionWidth: other.fractionWidth,
             signed: s)
         .widen(other)
         .value;
-    return FixedPointValue.populator(
+    return FixedPointValue.populatorWithSignedness(
             integerWidth: mr, fractionWidth: nr, signed: s)
         .ofLogicValue(val1 * val2);
   }
@@ -346,11 +380,11 @@ class FixedPointValue implements Comparable<FixedPointValue> {
     final mr = m1 + other.fractionWidth;
     final nr = fractionWidth + m2;
     final tr = mr + nr;
-    var val1 = FixedPointValue.populator(
+    var val1 = FixedPointValue.populatorWithSignedness(
             integerWidth: m1, fractionWidth: tr - m1, signed: s)
         .widen(this)
         .value;
-    var val2 = FixedPointValue.populator(
+    var val2 = FixedPointValue.populatorWithSignedness(
             integerWidth: tr - other.fractionWidth,
             fractionWidth: other.fractionWidth,
             signed: s)
@@ -370,7 +404,7 @@ class FixedPointValue implements Comparable<FixedPointValue> {
     if (isNegative() != other.isNegative()) {
       val = (~val) + 1;
     }
-    return FixedPointValue.populator(
+    return FixedPointValue.populatorWithSignedness(
             integerWidth: mr, fractionWidth: nr, signed: s)
         .ofLogicValue(val);
   }